fix bugs and move buffer helper methods into separate class
