+++ examples/main.js
@@ -65,7 +65,7 @@ function navigate () {
     lastEvalFn = fn.toString();
     sourceEl.textContent = unwrap(lastEvalFn);
 
-    // hljs.highlightBlock(sourceEl);
+    hljs.highlightBlock(sourceEl);
     bap.clock.stop();
     fn();
